Added All-Out, replaces All-In
All-Out is a build mission using Breakout as a base.  This is an aggressive mission where the player takes on the role Raynor traditionally plays in Breakout, with significantly fewer resources and no Tosh hero unit to assist him.  Enemy layouts have been mostly unchanged, and attack waves will arrive at the same rate as they did in the original mission.